cmd/sst-ctl: refactor
Split out subcommands to separate files.
